// src/app/model/producerDetails/producer-details.ts
export class ProducerDetails {
  id?: string;
  nome?: string;
  email?: string;
  products?: string;
  province?: string;
  city?: string;
  address?: string;
  phone?: string;
  description?: string;
  coverPhoto?: any;
  linkFacebook?: string;
  linkInstagram?: string;
  linkWebsite?: string;
  urlPhoto?: any;
}